+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - check-headers.sh file - corrected the path of check-ms-license.pl
 - removed unused doc_snippets
 - removed unneeded '*' at start of expressions in utils/check_license/check-headers.sh
+- (examples) use HELLO_STR_SIZE instead of KILOBYTE in case of the hello string
 
 ### Changed
 - logging of the source and the destination GID addresses in rpma_conn_req_new_from_id()